+++ awslimitchecker/services/cloudtrail.py
@@ -1,3 +1,43 @@
+"""
+awslimitchecker/services/cloudtrail.py
+
+The latest version of this package is available at:
+<https://github.com/jantman/awslimitchecker>
+
+################################################################################
+Copyright 2015-2018 Jason Antman <jason@jasonantman.com>
+
+    This file is part of awslimitchecker, also known as awslimitchecker.
+
+    awslimitchecker is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    awslimitchecker is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with awslimitchecker.  If not, see <http://www.gnu.org/licenses/>.
+
+The Copyright and Authors attributions contained herein may not be removed or
+otherwise altered, except to add the Author attribution of a contributor to
+this work. (Additional Terms pursuant to Section 7b of the AGPL v3)
+################################################################################
+While not legally required, I sincerely request that anyone who finds
+bugs please submit them at <https://github.com/jantman/awslimitchecker> or
+to me via email, and that you send any contributions or improvements
+either as a pull request on GitHub, or to me via email.
+################################################################################
+
+AUTHORS:
+Jason Antman <jason@jasonantman.com> <http://www.jasonantman.com>
+################################################################################
+"""
+
+import abc  # noqa
 import logging
 
 from .base import _AwsService
@@ -7,6 +47,7 @@ logger = logging.getLogger(__name__)
 
 
 class _CloudTrailService(_AwsService):
+
     service_name = 'CloudTrail'
     api_name = 'cloudtrail'
     aws_type = 'AWS::CloudTrail::Trail'
@@ -40,18 +81,20 @@ class _CloudTrailService(_AwsService):
             event_selectors = response['EventSelectors']
 
             for event_selector in event_selectors:
-                data_resources = event_selector['DataResources']
-                data_resource_count += len(event_selector['DataResources']) \
-                    if data_resources else 0
+                data_resource_count += len(
+                    event_selector.get('DataResources', [])
+                )
 
             self.limits['Event Selectors Per Trail']._add_current_usage(
                 len(event_selectors),
-                aws_type=self.aws_type
+                aws_type='AWS::CloudTrail::EventSelector',
+                resource_id=trail['Name']
             )
 
             self.limits['Data Resources Per Trail']._add_current_usage(
                 data_resource_count,
-                aws_type=self.aws_type
+                aws_type='AWS::CloudTrail::DataResource',
+                resource_id=trail['Name']
             )
 
         self.limits['Trails Per Region']._add_current_usage(
@@ -90,7 +133,8 @@ class _CloudTrailService(_AwsService):
             5,
             self.warning_threshold,
             self.critical_threshold,
-            limit_type=self.aws_type
+            limit_type=self.aws_type,
+            limit_subtype='AWS::CloudTrail::EventSelector'
         )
 
         limits['Data Resources Per Trail'] = AwsLimit(
@@ -99,7 +143,8 @@ class _CloudTrailService(_AwsService):
             250,
             self.warning_threshold,
             self.critical_threshold,
-            limit_type=self.aws_type
+            limit_type=self.aws_type,
+            limit_subtype='AWS::CloudTrail::DataResource'
         )
 
         self.limits = limits
@@ -116,5 +161,5 @@ class _CloudTrailService(_AwsService):
         """
         return [
             "cloudtrail:DescribeTrails",
-            "cloudtrail:GetTrailStatus",
+            "cloudtrail:GetEventSelectors",
         ]