+++ awslimitchecker/tests/services/test_cloudtrail.py
@@ -1,3 +1,42 @@
+"""
+awslimitchecker/services/cloudtrail.py
+
+The latest version of this package is available at:
+<https://github.com/jantman/awslimitchecker>
+
+################################################################################
+Copyright 2015-2018 Jason Antman <jason@jasonantman.com>
+
+    This file is part of awslimitchecker, also known as awslimitchecker.
+
+    awslimitchecker is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    awslimitchecker is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with awslimitchecker.  If not, see <http://www.gnu.org/licenses/>.
+
+The Copyright and Authors attributions contained herein may not be removed or
+otherwise altered, except to add the Author attribution of a contributor to
+this work. (Additional Terms pursuant to Section 7b of the AGPL v3)
+################################################################################
+While not legally required, I sincerely request that anyone who finds
+bugs please submit them at <https://github.com/jantman/awslimitchecker> or
+to me via email, and that you send any contributions or improvements
+either as a pull request on GitHub, or to me via email.
+################################################################################
+
+AUTHORS:
+Jason Antman <jason@jasonantman.com> <http://www.jasonantman.com>
+################################################################################
+"""
+
 import sys
 
 from awslimitchecker.tests.services import result_fixtures
@@ -47,10 +86,14 @@ class Test_CloudTrailService(object):
 
         event_selectors_limit = limit_dict['Event Selectors Per Trail']
         assert event_selectors_limit.limit_type == AWS_TYPE
+        assert event_selectors_limit.limit_subtype == \
+            'AWS::CloudTrail::EventSelector'
         assert event_selectors_limit.default_limit == 5
 
         data_resources_limit = limit_dict['Data Resources Per Trail']
         assert data_resources_limit.limit_type == AWS_TYPE
+        assert data_resources_limit.limit_subtype == \
+            'AWS::CloudTrail::DataResource'
         assert data_resources_limit.default_limit == 250
 
     def test_get_limits_again(self):
@@ -65,8 +108,16 @@ class Test_CloudTrailService(object):
         mock_trails = Mock()
         mock_trails.describe_trails.return_value = \
             result_fixtures.CloudTrail.mock_describe_trails
-        mock_trails.get_event_selectors.return_value = \
-            result_fixtures.CloudTrail.mock_get_event_selectors
+
+        def se_selectors(*args, **kwargs):
+            if kwargs['TrailName'] == 'trail2':
+                return result_fixtures.CloudTrail.mock_get_event_selectors
+            return {
+                'TrailArn': 'arn:%s' % kwargs['TrailName'],
+                'EventSelectors': []
+            }
+
+        mock_trails.get_event_selectors.side_effect = se_selectors
 
         with patch('%s.connect' % PATCH_BASE,) as mock_connect:
             cls = _CloudTrailService(21, 43)
@@ -83,19 +134,19 @@ class Test_CloudTrailService(object):
 
         usage = cls.limits['Event Selectors Per Trail'].get_current_usage()
         assert len(usage) == 3
-        assert usage[0].get_value() == 1
-        assert usage[1].get_value() == 1
-        assert usage[2].get_value() == 1
+        assert usage[0].get_value() == 0
+        assert usage[1].get_value() == 3
+        assert usage[2].get_value() == 0
 
         usage = cls.limits['Data Resources Per Trail'].get_current_usage()
         assert len(usage) == 3
-        assert usage[0].get_value() == 3
+        assert usage[0].get_value() == 0
         assert usage[1].get_value() == 3
-        assert usage[2].get_value() == 3
+        assert usage[2].get_value() == 0
 
     def test_required_iam_permissions(self):
         cls = _CloudTrailService(21, 43)
         assert cls.required_iam_permissions() == [
             "cloudtrail:DescribeTrails",
-            "cloudtrail:GetTrailStatus",
+            "cloudtrail:GetEventSelectors",
         ]